updated data generator
